Validate topped in topic

+++ src/shared/base.rs
@@ -1,5 +1,5 @@
 use chrono::prelude::Utc;
 
-pub fn timestamp() -> u64 {
-    return Utc::now().timestamp() as u64;
+pub fn timestamp() -> i64 {
+    return Utc::now().timestamp();
 }
